Practical 1

+++ function-1-2.cpp
@@ -2,12 +2,12 @@ double array_mean(int array[], int n) {
     if (n < 1) {
     return 0.0;
   }
-    int sum = 0;
+    double sum = 0;
     double mean;
     for (int i = 0; i < n; i++) {
         sum = sum + array[i];
     }
-    mean = double(sum /n);
+    mean = sum / static_cast<double>(n);
     return mean;
 
 }
